+++ lib/Elastica/ResultSet.php
@@ -101,7 +101,7 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     {
         $data = $this->_response->getData();
 
-        return isset($data['suggest']) ? $data['suggest'] : [];
+        return $data['suggest'] ?? [];
     }
 
     /**
@@ -125,7 +125,7 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     {
         $data = $this->_response->getData();
 
-        return isset($data['aggregations']) ? $data['aggregations'] : [];
+        return $data['aggregations'] ?? [];
     }
 
     /**
@@ -156,7 +156,7 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     {
         $data = $this->_response->getData();
 
-        return isset($data['hits']['total']) ? (int) $data['hits']['total'] : 0;
+        return (int) ($data['hits']['total'] ?? 0);
     }
 
     /**
@@ -168,7 +168,7 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     {
         $data = $this->_response->getData();
 
-        return isset($data['hits']['max_score']) ? (float) $data['hits']['max_score'] : 0;
+        return (float) ($data['hits']['max_score'] ?? 0);
     }
 
     /**
@@ -180,7 +180,7 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     {
         $data = $this->_response->getData();
 
-        return isset($data['took']) ? $data['took'] : 0;
+        return $data['took'] ?? 0;
     }
 
     /**
@@ -236,15 +236,11 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     /**
      * Returns the current object of the set.
      *
-     * @return \Elastica\Result|bool Set object or false if not valid (no more entries)
+     * @return \Elastica\Result Set object
      */
     public function current()
     {
-        if ($this->valid()) {
-            return $this->_results[$this->key()];
-        } else {
-            return false;
-        }
+        return $this->_results[$this->key()];
     }
 
     /**
@@ -253,8 +249,6 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     public function next()
     {
         ++$this->_position;
-
-        return $this->current();
     }
 
     /**
@@ -308,15 +302,15 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
      *
      * @throws Exception\InvalidException If offset doesn't exist
      *
-     * @return Result|null
+     * @return Result
      */
     public function offsetGet($offset)
     {
         if ($this->offsetExists($offset)) {
             return $this->_results[$offset];
-        } else {
-            throw new InvalidException('Offset does not exist.');
         }
+
+        throw new InvalidException('Offset does not exist.');
     }
 
     /**