+++ lib/Elastica/Script/AbstractScript.php
@@ -56,16 +56,16 @@ abstract class AbstractScript extends AbstractUpdateAction
 
     private static function _createFromArray(array $data)
     {
-        $params = isset($data['script']['params']) ? $data['script']['params'] : [];
-        $lang = isset($data['script']['lang']) ? $data['script']['lang'] : null;
+        $params = $data['script']['params'] ?? [];
+        $lang = $data['script']['lang'] ?? null;
 
         if (!is_array($params)) {
             throw new InvalidException('Script params must be an array');
         }
 
-        if (isset($data['script']['inline'])) {
+        if (isset($data['script']['source'])) {
             return new Script(
-                $data['script']['inline'],
+                $data['script']['source'],
                 $params,
                 $lang
             );