+++ test/Elastica/ClientTest.php
@@ -342,8 +342,8 @@ class ClientTest extends BaseTest
         // deleteIds are the type we are testing for
         $idxString = $index->getName();
         $typeString = $type->getName();
-        $this->assertTrue(is_string($idxString));
-        $this->assertTrue(is_string($typeString));
+        $this->assertInternalType('string', $idxString);
+        $this->assertInternalType('string', $typeString);
 
         // Try to delete doc with a routing value which hashes to
         // a different shard then the id.
@@ -418,7 +418,7 @@ class ClientTest extends BaseTest
         // And verify that the variables we are doing to send to
         // deleteIds are the type we are testing for
         $idxString = $index->getName();
-        $this->assertTrue(is_string($idxString));
+        $this->assertInternalType('string', $idxString);
         $this->assertInstanceOf(Type::class, $type);
 
         // Using the existing $index and $type variables which
@@ -483,7 +483,7 @@ class ClientTest extends BaseTest
         // deleteIds are the type we are testing for
         $typeString = $type->getName();
         $this->assertInstanceOf(Index::class, $index);
-        $this->assertTrue(is_string($typeString));
+        $this->assertInternalType('string', $typeString);
 
         // Using the existing $index and $type variables which
         // are \Elastica\Index and \Elastica\Type objects respectively
@@ -579,7 +579,7 @@ class ClientTest extends BaseTest
         $connections = $client->getConnections();
 
         // two connections are setup
-        $this->assertEquals(2, count($connections));
+        $this->assertCount(2, $connections);
 
         // One connection has to be disabled
         $this->assertTrue($connections[0]->isEnabled() == false || $connections[1]->isEnabled() == false);
@@ -607,7 +607,7 @@ class ClientTest extends BaseTest
         $connections = $client->getConnections();
 
         // two connections are setup
-        $this->assertEquals(2, count($connections));
+        $this->assertCount(2, $connections);
 
         // One connection has to be disabled
         $this->assertTrue($connections[0]->isEnabled() == false || $connections[1]->isEnabled() == false);
@@ -900,7 +900,7 @@ class ClientTest extends BaseTest
         $response = $client->addDocuments($docs);
 
         $this->assertInstanceOf(ResponseSet::class, $response);
-        $this->assertEquals(3, count($response));
+        $this->assertCount(3, $response);
         $this->assertTrue($response->isOk());
         $this->assertFalse($response->hasError());
         $this->assertEquals('', $response->getError());
@@ -917,12 +917,55 @@ class ClientTest extends BaseTest
         $response = $client->deleteDocuments($deleteDocs);
 
         $this->assertInstanceOf(ResponseSet::class, $response);
-        $this->assertEquals(2, count($response));
+        $this->assertCount(2, $response);
         $this->assertTrue($response->isOk());
         $this->assertFalse($response->hasError());
         $this->assertEquals('', $response->getError());
 
         $index->refresh();
+
+        $this->assertEquals(1, $type->count());
+    }
+
+    /**
+     * @group functional
+     */
+    public function testDeleteDocumentsWithRequestParameters()
+    {
+        $index = $this->_createIndex();
+        $type = $index->getType('test');
+        $client = $index->getClient();
+
+        $docs = [
+            new Document(1, ['field' => 'value1'], $type, $index),
+            new Document(2, ['field' => 'value2'], $type, $index),
+            new Document(3, ['field' => 'value3'], $type, $index),
+        ];
+
+        $response = $client->addDocuments($docs);
+
+        $this->assertInstanceOf(ResponseSet::class, $response);
+        $this->assertCount(3, $response);
+        $this->assertTrue($response->isOk());
+        $this->assertFalse($response->hasError());
+        $this->assertEquals('', $response->getError());
+
+        $index->refresh();
+
+        $this->assertEquals(3, $type->count());
+
+        $deleteDocs = [
+            $docs[0],
+            $docs[2],
+        ];
+
+        $response = $client->deleteDocuments($deleteDocs, ['refresh' => true]);
+
+        $this->assertInstanceOf(ResponseSet::class, $response);
+        $this->assertCount(2, $response);
+        $this->assertTrue($response->isOk());
+        $this->assertFalse($response->hasError());
+        $this->assertEquals('', $response->getError());
 
         $this->assertEquals(1, $type->count());
     }
@@ -1032,6 +1075,34 @@ class ClientTest extends BaseTest
             $this->assertTrue($doc->hasId());
             $this->assertTrue($doc->hasVersion());
             $this->assertEquals(1, $doc->getVersion());
+        }
+    }
+
+    /**
+     * @group functional
+     */
+    public function testAddDocumentsPipeline()
+    {
+        $docs = [];
+        for ($i = 0; $i < 10; ++$i) {
+            $docs[] = new Document(null, ['old' => $i]);
+        }
+
+        $index = $this->_createIndex();
+        $this->_createRenamePipeline();
+
+        $client = $index->getClient();
+        $client->setConfigValue('document', ['autoPopulate' => true]);
+
+        $type = $index->getType('test');
+        $type->addDocuments($docs, ['pipeline' => 'renaming']);
+
+        foreach ($docs as $i => $doc) {
+            $foundDoc = $type->getDocument($doc->getId());
+            $this->assertInstanceOf(Document::class, $foundDoc);
+            $data = $foundDoc->getData();
+            $this->assertArrayHasKey('new', $data);
+            $this->assertEquals($i, $data['new']);
         }
     }
 
@@ -1321,7 +1392,7 @@ class ClientTest extends BaseTest
         $endpoint->setParams(['types' => [$type->getName()]]);
         $response = $client->requestEndpoint($endpoint);
 
-        $this->assertTrue(isset($response->getData()['indices'][$index->getName()]['total']['indexing']['types']));
+        $this->assertArrayHasKey('types', $response->getData()['indices'][$index->getName()]['total']['indexing']);
 
         $this->assertEquals(
             ['test'],