+++ test/Elastica/Query/CommonTest.php
@@ -52,12 +52,60 @@ class CommonTest extends BaseTest
         $results = $type->search($query)->getResults();
 
         //documents containing only common words should not be returned
-        $this->assertEquals(3, sizeof($results));
+        $this->assertCount(3, $results);
 
         $query->setMinimumShouldMatch(2);
         $results = $type->search($query);
 
         //only the document containing both low frequency terms should match
         $this->assertEquals(1, $results->count());
+    }
+
+    /**
+     * @group unit
+     */
+    public function testSetHighFrequencyOperator()
+    {
+        $value = 'OPERATOR_TEST';
+        $query = new Common('body', 'test query', .001);
+        $query->setHighFrequencyOperator($value);
+
+        $this->assertEquals($value, $query->toArray()['common']['body']['high_frequency_operator']);
+    }
+
+    /**
+     * @group unit
+     */
+    public function testSetBoost()
+    {
+        $value = .02;
+        $query = new Common('body', 'test query', .001);
+        $query->setBoost($value);
+
+        $this->assertEquals($value, $query->toArray()['common']['body']['boost']);
+    }
+
+    /**
+     * @group
+     */
+    public function testSetAnalyzer()
+    {
+        $value = 'test';
+        $query = new Common('body', 'test query', .001);
+        $query->setBoost($value);
+
+        $this->assertEquals($value, $query->toArray()['common']['body']['analyzer']);
+    }
+
+    /**
+     * @group
+     */
+    public function testSetDisableCoord()
+    {
+        $value = true;
+        $query = new Common('body', 'test query', .001);
+        $query->setDisableCoord($value);
+
+        $this->assertEquals($value, $query->toArray()['common']['body']['disable_coord']);
     }
 }